TextToSQL/requirements.txt
openai
pandas
numpy
streamlit
google-generativeai
python-dotenvs